requirements.txt
Django==3.2.15
mysqlclient==2.1.1
djangorestframework==3.13.1
djangorestframework-jwt==1.11.0
django-filter
django-cors-headers
pillow

django-paypal
requests